Fix logic for accepting packets overlapping rcv.nxt.
A packet that had data starting before rcv.nxt and ending right before
rcv.nxt was considered acceptable, but was reduced to length 0. Later on,
this caused utcp_recv() to return without sending an ACK packet back,
therefore the sender could think the packet was dropped and would try to
retransmit the exact same packet over and over.

The fix is to treat these packets as unacceptable, which causes an ACK to
be returned informing the sender that we already have this data.