Optimize many calls to utcp_send() with small amount of data.
When sending a lot of data in small chunks, it can happen that UTCP
never tries to coalesce data into MTU-sized packets. Help by not calling
ack() when we send less than the MTU and the amount of unacknowledged
data is more than the MTU. This causes subsequent utcp_send() calls to
add more data to the buffer without sending small packets. UTCP will
start sending the queued data when ACK packets arrive for the
unacknowledged data. Think of it as the Nagle algorithm, except that
it's disabled for the first MTU of data.

This is not the full solution to the problem, the ACK handling code
should also be taught to avoid sending small packets.